plugins: merge MoreQuickReactions into betterQuickReact
